[Bronze I] Title: 피보나치 수 2, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/2748. 피보나치 수 2/피보나치 수 2.c
@@ -2,9 +2,9 @@
 int main(void){
     int num;
     scanf("%d", &num);
-    long long arr[90] = {0, 1};
+    long arr[90] = {0, 1};
     for (int i=2;i<=num;i++)
         arr[i] = arr[i-1] + arr[i-2];
-    printf("%lld", arr[num]);
+    printf("%ld", arr[num]);
     return 0;
 }
